API/feowl-api/requirements.txt
Django==1.4
South==0.7.5
Werkzeug==0.8.3
argparse==1.2.1
django-extensions==0.9
django-grappelli==2.3.8
django-rosetta==0.6.8
django-tastypie==0.9.11
django-templatesadmin==0.6
gunicorn==0.14.5
mimeparse==0.1.3
numpy==1.6.2
psycopg2==2.4.5
python-dateutil==1.5
pytz==2012c
wsgiref==0.1.2
